VBox/com: Added SafeArray::push_front() and implemented a new (initial) testcase for the SafeArray glue code [build fix].
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@99771 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/testcase/tstSafeArray.cpp
@@ -59,7 +59,7 @@ int main()
         RTTESTI_CHECK(aInt2.size() == 42);
         aInt2.setNull();
 
-        com::SafeArray<int> aInt3(0);
+        com::SafeArray<int> aInt3((size_t)0);
         RTTESTI_CHECK(aInt3.size() == 0);
         aInt3.setNull();
         RTTESTI_CHECK(aInt3.size() == 0);
